// src/model/LoggedRequest.ts
/**
 * This file was automatically generated. DO NOT MODIFY IT BY HAND.
 */

export interface LoggedRequest {
  /**
   * The HTTP request method
   */
  method?: string;
  /**
   * The path and query to match exactly against
   */
  url?: string;
  /**
   * The full URL to match against
   */
  absoluteUrl?: string;
  /**
   * Header patterns to match against in the <key>: { "<predicate>": "<value>" } form
   */
  headers?: {
    [k: string]: any;
  };
  /**
   * Cookie patterns to match against in the <key>: { "<predicate>": "<value>" } form
   */
  cookies?: {
    [k: string]: any;
  };
  /**
   * Body string to match against
   */
  body?: string;
  [k: string]: any;
}